frontend/requirements.txt
streamlit==1.21.0
pandas==1.5.3
requests==2.28.2
pyyaml==6.0
joblib==1.2.0
numpy==1.24.2
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
scikit-learn==1.2.2
lightgbm==3.3.5
optuna==3.1.1
jsonschema==4.17.3
